chore: move UserAvatar to kit directory
