+++ webui/react/src/components/ProjectCard.tsx
@@ -2,9 +2,9 @@ import { Typography } from 'antd';
 import React, { useCallback } from 'react';
 
 import Tooltip from 'components/kit/Tooltip';
+import Avatar from 'components/kit/UserAvatar';
 import Link from 'components/Link';
 import TimeAgo from 'components/TimeAgo';
-import Avatar from 'components/UserAvatar';
 import { paths } from 'routes/utils';
 import Icon from 'shared/components/Icon/Icon';
 import { routeToReactUrl } from 'shared/utils/routes';
