+++ webui/react/src/components/kit/UserAvatar.test.tsx
