+++ webui/react/src/pages/WorkspaceList/WorkspaceCard.tsx
@@ -4,8 +4,8 @@ import React, { useCallback } from 'react';
 
 import DynamicIcon from 'components/DynamicIcon';
 import Tooltip from 'components/kit/Tooltip';
+import Avatar from 'components/kit/UserAvatar';
 import Link from 'components/Link';
-import Avatar from 'components/UserAvatar';
 import { paths } from 'routes/utils';
 import Icon from 'shared/components/Icon/Icon';
 import { routeToReactUrl } from 'shared/utils/routes';
